split component class into partial classes
